# Cache reserved keywords