--- a/naga/src/back/hlsl/keywords.rs
+++ b/naga/src/back/hlsl/keywords.rs
@@ -1,3 +1,7 @@
+use std::sync::LazyLock;
+
+use hashbrown::HashSet;
+
 // When compiling with FXC without strict mode, these keywords are actually case insensitive.
 // If you compile with strict mode and specify a different casing like "Pass" instead in an identifier, FXC will give this error:
 // "error X3086: alternate cases for 'pass' are deprecated in strict mode"
@@ -911,6 +915,22 @@ pub const TYPES: &[&str] = &{
 
     res
 };
+
+/// The above set of reserved keywords, turned into a cached HashSet. This saves
+/// significant time during [`Namer::reset`](crate::proc::Namer::reset).
+///
+/// See <https://github.com/gfx-rs/wgpu/pull/7338> for benchmarks.
+pub static RESERVED_SET: LazyLock<HashSet<&'static str>> = LazyLock::new(|| {
+    let mut set = HashSet::default();
+    set.reserve(RESERVED.len() + TYPES.len());
+    for &word in RESERVED {
+        set.insert(word);
+    }
+    for &word in TYPES {
+        set.insert(word);
+    }
+    set
+});
 
 pub const RESERVED_PREFIXES: &[&str] = &[
     "__dynamic_buffer_offsets",