**Description**

Found an interesting and easy to fix issue when fooling around with profiling our benchmarks. The text based backends spend an outsided amount of time on a single line of code every single time - calling `extend` on the reserved keyword hashmap.

This PR caches that hashmap in a LazyLock, so that you only need to build that hashmap once. We might look into `phf` at some point, but that's a bigger question.

Numbers from running our benchmark suite on a 9800x3d.

| Backend | Before | After | Difference |
|--------|--------|--------|--------|
| wgsl-out | 14.3µs | 10.4µs | **1.38x** |
| msl-out | 21.0µs | 14.6µs | **1.44x** |
| hlsl-out | 54.7µs | 16.4µs | **3.33x** |
| glsl-out | 9.8µs | 5.4µs | **1.81x** |

**Squash or Rebase?**

Squash