@cwfitzgerald 2025-03-14
cc @kpreid as you're pushing no-std. This kinda regresses us on the no-std side of things, and I don't know what to do about it. Maybe long term it's switching to `phf` or such, but we should have a strategy for lacylocks.