@kpreid 2025-03-14
By total coincidence, I was just today looking at this code, and I have a suggestion for an alternative: Sort the keyword lists, then use a binary search for lookups (preferably wrapped up into a nice `struct KeywordSet(&'static [&'static str])` with documented sortedness invariant and `get()` function).

Further ideas:

* Maybe record the maximum length of any keyword, and don't search at all if the identifier is longer? The max length for most languages is pretty long, though, so this filter won’t catch much of the most common variables.
* Use `phf`, as you said.
* Or some other build-time preprocessing. Generating a string matching automaton in Rust code or bytecode?

Also, I'd be interested to know how these various choices affect the binary size, not just run time. That matters for all those web applications that are including Naga in Wasm code.