@kpreid 2025-03-14
As to `LazyLock`s in general, my strategy in all cases like this is to figure out some structure for the data that can be generated in `build.rs` (or a macro, but that hasn’t been the case anywhere yet) and used as static data.