@cwfitzgerald 2025-03-15
Unrelated but for some reason this post is looooooong on github mobile. 

![Screenshot_20250314_202451_GitHub.jpg](https://github.com/user-attachments/assets/c985aac8-a330-429d-bd48-f7c7c1f538b9)