@waywardmonkeys 2025-03-15
You might take a look at the code that @raphlinus wrote in the `color` crate that does a light weight perfect hash function: https://github.com/linebender/color/blob/main/color/make_x11_colors.py